refactor: LottoConfig -> LottoConstant
