+++ src/main/kotlin/domain/check/LottoNumbersCheck.kt
@@ -1,12 +1,12 @@
 package org.example.domain.check
 
-import org.example.config.LottoConfig
+import org.example.constant.LottoConstant
 
 @JvmInline
 value class LottoNumbersCheck(private val numbers: List<Int>) {
     init {
-        require(numbers.size == LottoConfig.LOTTO_TICKET_SIZE) { LottoConfig.ERROR_MESSAGE_LOTTO_SIZE }
-        require(numbers.distinct().size == LottoConfig.LOTTO_TICKET_SIZE) { LottoConfig.ERROR_MESSAGE_DUPLICATE }
-        require(numbers.all { it in (LottoConfig.START_NUMBER..LottoConfig.END_NUMBER) }) { LottoConfig.ERROR_MESSAGE_NUMBER_RANGE }
+        require(numbers.size == LottoConstant.LOTTO_TICKET_SIZE) { LottoConstant.ERROR_MESSAGE_LOTTO_SIZE }
+        require(numbers.distinct().size == LottoConstant.LOTTO_TICKET_SIZE) { LottoConstant.ERROR_MESSAGE_DUPLICATE }
+        require(numbers.all { it in (LottoConstant.START_NUMBER..LottoConstant.END_NUMBER) }) { LottoConstant.ERROR_MESSAGE_NUMBER_RANGE }
     }
 }
