+++ src/main/kotlin/domain/generator/AutoLottoGenerator.kt
@@ -1,12 +1,12 @@
 package org.example.domain.generator
 
-import org.example.config.LottoConfig
+import org.example.constant.LottoConstant
 import org.example.domain.LottoTicket
 import org.example.domain.WinningLotto
 
 class AutoLottoGenerator() : LottoGenerator {
-    private val numberRange = LottoConfig.START_NUMBER..LottoConfig.END_NUMBER
-    private val numberCount = LottoConfig.LOTTO_TICKET_SIZE
+    private val numberRange = LottoConstant.START_NUMBER..LottoConstant.END_NUMBER
+    private val numberCount = LottoConstant.LOTTO_TICKET_SIZE
 
     override fun generate(): LottoTicket {
         val numbers = numberRange.shuffled().take(numberCount).sorted()
